Use the other version of endpoints.service.ts
It seems as though the original demo used one version of the endpoints
file when the ui was running on node, and a different one when running
on tomcat. Running on Liberty seems to require the one from the old node
setup.
It isn't obvious why, but this seems to work.