Remove tests temporarily
The tests for the web application don't currently run in the build, I'm
not sure exactly why, and they are not running on liberty anyway. Remove
them for the moment, aim to put them back but running on liberty in the
future.
The tests for the schedule project are running on tomcat with
arquillian. Again remove them for the moment, until they can run on
liberty.